añadiendo funcionalidades a publicacion, corrigiendo apis y más

+++ publicacion_app/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render
-from rest_framework.decorators import api_view
+from rest_framework.decorators import api_view, permission_classes
+from rest_framework.permissions import IsAuthenticated
 from rest_framework.response import Response
 from rest_framework import status
 from .models import Publicacion
@@ -8,13 +9,14 @@ from .serializers import PublicacionSerializer
 
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated]) 
 def crear_publicacion(request):
     if request.method == 'POST':
         data = request.data
         serializer = PublicacionSerializer(data=data)
         
         if serializer.is_valid():
-            serializer.save()
+            serializer.save(usuario=request.user)
             return Response(serializer.data, status=status.HTTP_201_CREATED)
         
         return Response(serializer.errors, status=status.HTTP_400_BAD_REQUEST)
@@ -22,21 +24,25 @@ def crear_publicacion(request):
 
 
 @api_view(['POST'])
+@permission_classes([IsAuthenticated]) 
 def eliminar_publicacion(request):
     try:
         publicacion_id = request.data.get('id')
         if publicacion_id is None:
             return Response(status=status.HTTP_400_BAD_REQUEST)
         
+        try:
+            publicacion = Publicacion.objects.get(pk=publicacion_id)
+        except Publicacion.DoesNotExist:
+            return Response(status=status.HTTP_404_NOT_FOUND)
         
-        publicacion = Publicacion.objects.get(pk=publicacion_id)
-        publicacion.delete()
-        
-        return Response(status=status.HTTP_204_NO_CONTENT)
+
+        if publicacion.usuario == request.user:
+            publicacion.delete()
+            return Response({'mensaje':'publicación eliminada'}, status=status.HTTP_204_NO_CONTENT)
     
-    
-    except Publicacion.DoesNotExist:
-        return Response(status=status.HTTP_404_NOT_FOUND)
+    except Exception as e:
+        return Response({'error': str(e)}, status=status.HTTP_500_INTERNAL_SERVER_ERROR)
